+++ app/code/Darkwoolf/AskQuestion/Model/Question.php
@@ -3,6 +3,14 @@
 namespace Darkwoolf\AskQuestion\Model;
 
 use Darkwoolf\AskQuestion\Model\ResourceModel\Question as QuestionResource;
+use Magento\Framework\Model\Context;
+use Magento\Framework\Registry;
+use Magento\Store\Model\StoreManagerInterface;
+use Magento\Framework\Model\ResourceModel\AbstractResource;
+use Magento\Framework\Data\Collection\AbstractDb;
+use Magento\Framework\Model\AbstractModel;
+use Magento\Framework\Exception\NoSuchEntityException;
+
 
 /**
  * Class Question
@@ -39,29 +47,27 @@ class Question extends \Magento\Framework\Model\AbstractModel
     /** @var string  */
     const STATUS_SUCCESS = 'Success';
 
-    /**
-     * @var \Magento\Store\Model\StoreManagerInterface
-     */
+    /** @var StoreManagerInterface  */
     private $storeManager;
 
     /** @var string  */
     protected $_eventPrefix = 'darkwoolf_ask_question';
 
-
     /**
-     * @param \Magento\Framework\Model\Context $context
-     * @param \Magento\Framework\Registry $registry
-     * @param \Magento\Store\Model\StoreManagerInterface $storeManager
-     * @param \Magento\Framework\Model\ResourceModel\AbstractResource $resource
-     * @param \Magento\Framework\Data\Collection\AbstractDb $resourceCollection
+     * Question constructor.
+     * @param Context $context
+     * @param Registry $registry
+     * @param StoreManagerInterface $storeManager
+     * @param AbstractResource|null $resource
+     * @param AbstractDb|null $resourceCollection
      * @param array $data
      */
     public function __construct(
-        \Magento\Framework\Model\Context $context,
-        \Magento\Framework\Registry $registry,
-        \Magento\Store\Model\StoreManagerInterface $storeManager,
-        \Magento\Framework\Model\ResourceModel\AbstractResource $resource = null,
-        \Magento\Framework\Data\Collection\AbstractDb $resourceCollection = null,
+        Context $context,
+        Registry $registry,
+        StoreManagerInterface $storeManager,
+        AbstractResource $resource = null,
+        AbstractDb $resourceCollection = null,
         array $data = []
     ) {
         parent::__construct($context, $registry, $resource, $resourceCollection, $data);
@@ -88,8 +94,8 @@ class Question extends \Magento\Framework\Model\AbstractModel
     }
 
     /**
-     * @return \Magento\Framework\Model\AbstractModel
-     * @throws \Magento\Framework\Exception\NoSuchEntityException
+     * @return AbstractModel
+     * @throws NoSuchEntityException
      */
     public function beforeSave()
     {
