+++ app/code/Darkwoolf/AskQuestion/Model/ResourceModel/Question.php
@@ -2,11 +2,13 @@
 
 namespace Darkwoolf\AskQuestion\Model\ResourceModel;
 
+use Magento\Framework\Model\ResourceModel\Db\AbstractDb;
+
 /**
  * Class Question
- * @package Darkwoolf\AskQuestion\ResourceModel
+ * @package Darkwoolf\AskQuestion\Model\ResourceModel
  */
-class Question extends \Magento\Framework\Model\ResourceModel\Db\AbstractDb
+class Question extends AbstractDb
 {
     protected function _construct()
     {
