+++ app/code/Darkwoolf/AskQuestion/Setup/InstallSchema.php
@@ -4,8 +4,9 @@ namespace Darkwoolf\AskQuestion\Setup;
 
 use Magento\Framework\Setup\ModuleContextInterface;
 use Magento\Framework\Setup\SchemaSetupInterface;
+use Magento\Framework\Setup\InstallSchemaInterface;
 
-class InstallSchema implements \Magento\Framework\Setup\InstallSchemaInterface
+class InstallSchema implements InstallSchemaInterface
 {
 
     /**
