+++ app/code/Darkwoolf/AskQuestion/etc/frontend/di.xml
@@ -3,7 +3,7 @@
         xsi:noNamespaceSchemaLocation="urn:magento:framework:ObjectManager/etc/config.xsd">
     <type name="Darkwoolf\AskQuestion\Model\ResourceModel\AskQuestion\Collection">
         <plugin name="darkwoolf_askquestion_plugin"
-                type="Darkwoolf\AskQuestion\Plugin\AskQuestion\Model\ResourceModel\AskQuestion\Collection"
+                type="Darkwoolf\AskQuestion\Plugin\AskQuestion\Model\ResourceModel\Question\Collection"
         />
     </type>
 </config>
