+++ app/code/Darkwoolf/AskQuestion/etc/frontend/routes.xml
@@ -6,3 +6,4 @@
         </route>
     </router>
 </config>
+
