+++ app/code/Darkwoolf/AskQuestion/view/frontend/templates/product/view/ask_question.phtml
@@ -55,3 +55,4 @@
 </div>
 <hr>
 <?= $block->getChildHtml('askquestion.list') ?>
+
